Implemented ir-grid-mapping

+++ src/rgrid.c
@@ -39,10 +39,10 @@
 #include "rgrid.h"
 
 static void get_all_grid_addresses(int grid_address[][3], const int mesh[3]);
-static size_t get_grid_index_double_mesh(const int address_double[3],
-                                         const int mesh[3]);
-static size_t get_grid_index_single_mesh(const int address[3],
-                                         const int mesh[3]);
+static long get_grid_index_double_mesh(const int address_double[3],
+                                       const int mesh[3]);
+static long get_grid_index_single_mesh(const int address[3],
+                                       const int mesh[3]);
 static void reduce_grid_address(int address[3], const int mesh[3]);
 static void reduce_grid_address_double(int address[3], const int mesh[3]);
 
@@ -51,8 +51,8 @@ void rgd_get_all_grid_addresses(int grid_address[][3], const int mesh[3])
   get_all_grid_addresses(grid_address, mesh);
 }
 
-size_t rgd_get_grid_index_double_mesh(const int address_double[3],
-                                      const int mesh[3])
+long rgd_get_grid_index_double_mesh(const int address_double[3],
+                                    const int mesh[3])
 {
   return get_grid_index_double_mesh(address_double, mesh);
 }
@@ -73,7 +73,7 @@ void rgd_get_grid_address_double_mesh(int address_double[3],
 static void get_all_grid_addresses(int grid_address[][3], const int mesh[3])
 {
   int i, j, k;
-  size_t grid_index;
+  long grid_index;
   int address[3];
 
   for (i = 0; i < mesh[0]; i++) {
@@ -95,8 +95,8 @@ static void get_all_grid_addresses(int grid_address[][3], const int mesh[3])
   }
 }
 
-static size_t get_grid_index_double_mesh(const int address_double[3],
-                                         const int mesh[3])
+static long get_grid_index_double_mesh(const int address_double[3],
+                                       const int mesh[3])
 {
   int i;
   int address[3];
@@ -113,14 +113,14 @@ static size_t get_grid_index_double_mesh(const int address_double[3],
   return get_grid_index_single_mesh(address, mesh);
 }
 
-static size_t get_grid_index_single_mesh(const int address[3],
-                                         const int mesh[3])
+static long get_grid_index_single_mesh(const int address[3],
+                                       const int mesh[3])
 {
 #ifndef GRID_ORDER_XYZ
-  return (address[2] * mesh[0] * (size_t)(mesh[1])
+  return (address[2] * mesh[0] * (long)(mesh[1])
           + address[1] * mesh[0] + address[0]);
 #else
-  return (address[0] * mesh[1] * (size_t)(mesh[2])
+  return (address[0] * mesh[1] * (long)(mesh[2])
           + address[1] * mesh[2] + address[2]);
 #endif
 }
